Move out CheckBox to widget package and rename to AbstractCheckBox
